Trait pallet_deip::traits::DeipAssetSystem

pub trait DeipAssetSystem<AccountId> {
    type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy;
    type AssetId: Member + Parameter + Default + Copy + HasCompact;
    fn try_get_tokenized_project(id: &Self::AssetId) -> Option<ProjectId>;

    fn account_balance(
        account: &AccountId, 
        asset: &Self::AssetId
    ) -> Self::Balance;

    fn total_supply(asset: &Self::AssetId) -> Self::Balance;

    fn get_project_nfts(id: &ProjectId) -> Vec<Self::AssetId>;

    fn get_nft_balances(id: &Self::AssetId) -> Option<Vec<AccountId>>;

    fn transactionally_transfer(
        from: &AccountId, 
        asset: Self::AssetId, 
        transfers: &[(Self::Balance, AccountId)]
    ) -> Result<(), ()>;

    fn transactionally_reserve(
        account: &AccountId, 
        id: InvestmentId, 
        shares: &[(Self::AssetId, Self::Balance)], 
        asset: Self::AssetId
    ) -> Result<(), ReserveError<Self::AssetId>>;

    fn transactionally_unreserve(
        id: InvestmentId
    ) -> Result<(), UnreserveError<Self::AssetId>>;

    fn transfer_from_reserved(
        id: InvestmentId, 
        who: &AccountId, 
        asset: Self::AssetId, 
        amount: Self::Balance
    ) -> Result<(), UnreserveError<Self::AssetId>>;

    fn transfer_to_reserved(
        who: &AccountId, 
        id: InvestmentId, 
        amount: Self::Balance
    ) -> Result<(), UnreserveError<Self::AssetId>>;
}

Associated Types

type Balance: Member + Parameter + AtLeast32BitUnsigned + Default + Copy

The units in which asset balances are recorded.

type AssetId: Member + Parameter + Default + Copy + HasCompact

The arithmetic type of asset identifier.

Required methods

fn try_get_tokenized_project(id: &Self::AssetId) -> Option<ProjectId>

fn account_balance(account: &AccountId, asset: &Self::AssetId) -> Self::Balance

fn total_supply(asset: &Self::AssetId) -> Self::Balance

fn get_project_nfts(id: &ProjectId) -> Vec<Self::AssetId>

fn get_nft_balances(id: &Self::AssetId) -> Option<Vec<AccountId>>

fn transactionally_transfer(
from: &AccountId,
asset: Self::AssetId,
transfers: &[(Self::Balance, AccountId)]
) -> Result<(), ()>

fn transactionally_reserve(
account: &AccountId,
id: InvestmentId,
shares: &[(Self::AssetId, Self::Balance)],
asset: Self::AssetId
) -> Result<(), ReserveError<Self::AssetId>>

Tries to transfer assets specified by shares from account to a specific balance identified by id. Some collateral fee may be locked from account.

fn transactionally_unreserve(
id: InvestmentId
) -> Result<(), UnreserveError<Self::AssetId>>

Transfers all assets currently owned by id to the account, used in transactionally_reserve, in a transactional way.

fn transfer_from_reserved(
id: InvestmentId,
who: &AccountId,
asset: Self::AssetId,
amount: Self::Balance
) -> Result<(), UnreserveError<Self::AssetId>>

Transfers amount of assets id owned by account specified with id to who.

fn transfer_to_reserved(
who: &AccountId,
id: InvestmentId,
amount: Self::Balance
) -> Result<(), UnreserveError<Self::AssetId>>

Transfers amount of assets from who to account specified by id. Assets should be specified in call to transactionally_reserve.