Crate pallet_deip

Deip Module

A module for managing digital assets.

multisig::Config

Call

Overview

This module contains functionality for managing different types of digital assets.

It provides a hierarchy to simply operate digital assets in the real world. The module contains entities Project and Content of the Project with belongs to multi Account aka Team.

Besides, the Module provides Proof of share functionality. Proof of Share is a term we use for a special cryptographic proof that a sender actually sent, and the receiver has actually received an encrypted payload and a key to decrypt it. Please refer to the attached image. Includes entities like NDA and NDA Access Request.

Interface

Dispatchable Functions

add_domain - Add cryptographic hash of DomainId

create_project - Create Project belongs to Account (Team)

update_project - Update Project info

create_project_content - Create Project Content (Digital Asset)

create_project_nda - Create NDA contract between sides

create_nda_content_access_request - Some side request access to the data of contract

fulfill_nda_content_access_request - Granter fulfill access request to the data

reject_nda_content_access_request - Granter reject access request to the data

create_review - Create Review

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

Domain
PPossible project domains

GenesisConfig
Genesis config for the module, allow to build genesis storage.

H160
Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256
Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

Module

Nda
NDA contract between parties. Usually about dislocating or not dislocating some confidential info

NdaAccessRequest
NDA access request. One of the partice may decide to request to receive some info included into contract. Holder should fulfill or reject this request.

Project
Core entity of pallet. Everything connected to Project. Only Account (Team) stand before Project in hierarchy.

ProjectContent
Digital asset. Contains information of content and authors of Digital asset.

Review
Review

Enums

Call
Dispatchable calls.

Error

ProjectContentType
Possible statuses of Project inherited from Project Content type

RawEvent
Events for this module.

Constants

MAX_DOMAINS
A maximum number of Domains. When domains reaches this number, no new domains can be added.

Traits

Config
Configuration trait. Pallet depends on frame_system and pallet_timestamp.

Type Definitions

DomainId
Unique DomainId reference

Event
RawEvent specialized for the configuration Config

NdaAccessRequestId
Unique NdaAccess Request reference

NdaAccessRequestOf

NdaId
Unique NDA reference

NdaOf

ProjectContentId
Unique Project Contnt reference

ProjectContentOf

ProjectId
Unique Project ID reference

ProjectOf

ReviewId
Unique Review reference

ReviewOf